Refactor count header middleware
