+++ lib/tentd/api.rb
@@ -21,6 +21,7 @@ module TentD
     autoload :Router, 'tentd/api/router'
     autoload :Middleware, 'tentd/api/middleware'
     autoload :PaginationHeader, 'tentd/api/pagination_header'
+    autoload :CountHeader, 'tentd/api/count_header'
     include Router
 
     mount Apps
